+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ Changelog
 =======
 * Improves Container  admin queryset.
 * Redirects Container objects in admin to respective child_class edit page.
+* Refactoring related post to container migration
 
 0.2.8.3
 =======