+++ CHANGELOG.rst
@@ -4,6 +4,8 @@ Changelog
 
 0.2.8.4
 =======
+* Improves Container  admin queryset.
+* Redirects Container objects in admin to respective child_class edit page.
 * Refactoring related post to container migration
 
 0.2.8.3