+++ CHANGELOG.rst
@@ -2,8 +2,9 @@
 Changelog
 =========
 
-0.2.8.4
-=======
+0.2.9
+=====
+* QuerySet of containerbox select more one channel
 * Improves Container  admin queryset.
 * Redirects Container objects in admin to respective child_class edit page.
 * Refactoring related post to container migration